Allow setting styles for plots (see release notes).
