+++ andes/plot.py
@@ -10,10 +10,85 @@ import re
 import numpy as np
 
 from andes.core.var import BaseVar
-from andes.shared import plt, set_font, set_latex
+from andes.shared import find_executable, mpl, plt
+from andes.utils.paths import get_dot_andes_path  # NOQA
 
 logger = logging.getLogger(__name__)
-DPI = 80
+DPI = None
+
+
+def set_latex():
+    """
+    Enables LaTeX for matplotlib based on the `with_latex` option and `dvipng` availability.
+
+    Returns
+    -------
+    bool
+        True for LaTeX on, False for off
+    """
+
+    if find_executable('dvipng'):
+        mpl.rc('text', usetex=True)
+
+        no_warn_file = os.path.join(get_dot_andes_path(), '.no_warn_latex')
+        if not os.path.isfile(no_warn_file):
+            print('Using LaTeX for rendering. If an error occurs:')
+            print('a) If you are using `andes plot`, disable with option "-d",')
+            print('b) If you are using `plot()`, set "latex=False".')
+
+            try:
+                with open(os.path.join(get_dot_andes_path(), '.no_warn_latex'), 'w') as f:
+                    f.write('0')
+            except OSError:
+                pass
+
+        return True
+
+    return False
+
+
+def set_font(family='serif', size=12, style='normal', weight='normal'):
+    """
+    Sets the font for matplotlib.
+
+    Parameters
+    ----------
+    family : str
+        Font family.
+    size : int
+        Font size.
+    style : str
+        Font style.
+    weight : str
+        Font weight.
+    """
+
+    mpl.rc('font', family=family, size=size, style=style, weight=weight)
+
+
+def set_style(style='default'):
+    """
+    Set matplotlib style.
+
+    Parameters
+    ----------
+    style : str
+        `default`, `ieee` (require `scienceplots`), or other available styles
+        (see `matplotlib.pyplot.style.available`).
+    """
+
+    if style is None:
+        style = 'default'
+
+    if style == 'ieee':
+        try:
+            plt.style.use(['science', 'ieee'])
+        except (NameError, OSError):
+            logger.error("Please install `scienceplots` with `pip` to use the 'ieee' style.")
+
+    else:
+        plt.style.use(style)
+        set_font()
 
 
 class TDSData:
@@ -508,7 +583,7 @@ class TDSData:
                   hline1=None, hline2=None, vline1=None, hline=None, vline=None,
                   vline2=None, set_xlim=True, set_ylim=True, autoscale=False, figsize=None,
                   legend_bbox=None, legend_loc=None, legend_ncol=1,
-                  mask=True, color=None,
+                  mask=True, color=None, style='default',
                   **kwargs):
         """
         Plot lines for the supplied data and options.
@@ -549,7 +624,8 @@ class TDSData:
         >>>                      ylabel='Ipcmd [pu]')
 
         """
-        set_font(family='serif', size=font_size)
+
+        set_style(style)
 
         if not isinstance(ydata, np.ndarray):
             raise TypeError("ydata must be a numpy array. Retrieve with get_values().")
