+++ docs/source/release-notes.rst
@@ -12,6 +12,9 @@ v1.6 Notes
 v1.6.7 (2022-05-xx)
 -------------------
 - Fix the initialization of offline synchronous generators.
+- Allow styles to be set for plots using the argument ``style``. To generate
+  figures for IEEE publications, use ``style=ieee`` (require package
+  ``scienceplots``).
 
 v1.6.6 (2022-04-30)
 -------------------
